work on paimment journey and history

+++ src/app/journey/journey.component.ts
@@ -1,5 +1,6 @@
 import { FormGroup, FormBuilder, Validators } from "@angular/forms";
-import { UsersService } from "../api/services/users.service";
+import * as appHelpers from '../core/_helpers/app.helper';
+
 import {
   Component,
   OnInit,
@@ -7,11 +8,12 @@ import {
   AfterViewInit,
   OnDestroy
 } from "@angular/core";
-import { Subject, merge } from "rxjs";
+import { Subject, merge, of } from "rxjs";
 import { MatPaginator, MatSnackBar, MatDialog } from "@angular/material";
-import { startWith, tap, takeUntil, switchMap } from "rxjs/operators";
+import { startWith, tap, takeUntil, switchMap, catchError } from "rxjs/operators";
 import { JourneeDetailService, CaisseService } from "../api/services";
 import { ModifyJourneyComponent } from './Dialogs/modify-journey/modify-journey.component';
+import { ValidationService } from '../shared/services/validation.service';
 
 @Component({
   selector: "app-journey",
@@ -31,7 +33,6 @@ export class JourneyComponent implements OnInit, AfterViewInit, OnDestroy {
   caisseDetails;
   form: FormGroup;
   
-   numericNumberReg= '^-?[0-9]\\d*(\\.\\d{1,2})?$';
   constructor(
     fb: FormBuilder,
     public _snackBar: MatSnackBar,
@@ -41,15 +42,41 @@ export class JourneyComponent implements OnInit, AfterViewInit, OnDestroy {
 
   ) {
     this.form = fb.group({
-      montant: ["", [Validators.required,Validators.pattern(this.numericNumberReg)]],
-      motif: ["", [Validators.required]]
+      montant: ["", [Validators.required,ValidationService.NumberValidator,ValidationService.notNull]],
+      motif: ["", [Validators.required ]] 
     });
     
   }
 
   ngOnInit() {
     this.getCaisse();
-    this.getJourneys();
+  }
+  ngAfterViewInit(): void {
+    merge(this.paginator.page, this.search$)
+    .pipe(
+      startWith([{ pageIndex: 0, pageSize: 10 }, {}]),
+      takeUntil(this.unsubscribe$),
+      tap(() => {
+        this.isLoading = true;
+        this.error = undefined;
+        this.dataSource = { ...this.dataSource };
+      }),
+      switchMap(() =>
+      this.journeyService.GetAllJourneeDetail()
+          .pipe(
+            catchError(err => {
+              this.isLoading = false;
+              this.error = err;
+              return of(this.dataSource);
+            })
+          )
+      )
+    )
+    .subscribe((res: any) => {
+      this.isLoading = false;
+      this.dataSource = res;
+    });
+
   }
   editJourney(row) {
     this.dialog
@@ -66,30 +93,41 @@ export class JourneyComponent implements OnInit, AfterViewInit, OnDestroy {
     });
   }
   removeJourney(journey) {
-    this.journeyService.DeleteJourneeDetail(journey.idJourneeDetail).subscribe(res => {
-     if(res){
-      this.search$.next();
-      this.getCaisse();
-      this._snackBar.open("Journey deleted successfely", "x", {
-        duration: 3000,
-        panelClass: ["success-snackbar"]
-      });
-     }
+    appHelpers
+    .confirmDialog(
+      this.dialog,
+      'Do you really want to delete this line ?',
+      '',
+      'Yes',
+      'CANCEL'
+    )
+    .subscribe(value => {
+      if (value) {
+        this.journeyService.DeleteJourneeDetail(journey.idJourneeDetail).subscribe(res => {
+          if(res){
+           this.search$.next();
+           this.getCaisse();
+           this._snackBar.open("Journey deleted successfely", "x", {
+             duration: 3000,
+             panelClass: ["success-snackbar"]
+           });
+          }
+         });
+      }
     });
   }
-getCaisse(){
-  this.caisseService.GetCaisseEnCours().subscribe(res => {
-    if (res) {
-      console.log(res);
-      this.caisseDetails = res;
-    }
-  });
-}
+    getCaisse(){
+    this.caisseService.GetCaisseEnCours().subscribe(res => {
+      if (res) {
+        console.log(res);
+        this.caisseDetails = res;
+      }
+    });
+  }
   addJourney() {
-     console.log(this.form.value);
      if (this.form.valid) {
-      console.log(this.form.value);
-      this.journeyService
+      if(this.form.get('montant').value !=0){
+        this.journeyService
         .AddJourneeDetail({
           Prix: this.form.controls.montant.value,
           Motif: this.form.controls.motif.value,
@@ -98,12 +136,18 @@ getCaisse(){
           if (res) {
             this.getJourneys();
             this.getCaisse();
-            this._snackBar.open("Journey added successfely", "x", {
+            this._snackBar.open("Journey updated successfely", "x", {
               duration: 3000,
               panelClass: ["success-snackbar"]
             });
           }
         });
+      }else{
+        this._snackBar.open("amount shouldn't be null", "x", {
+          duration: 4000,
+          panelClass: ["danger-snackbar"]
+        });
+      }
       }else{
         this._snackBar.open("champs required", "x", {
           duration: 3000,
@@ -112,24 +156,10 @@ getCaisse(){
       }
   }
   
-
   getJourneys() {
-    merge(this.search$)
-      .pipe(
-        startWith([]),
-        takeUntil(this.unsubscribe$),
-        tap(() => {
-          this.isLoading = true;
-          this.error = undefined;
-        }),
-        switchMap(() => this.journeyService.GetAllJourneeDetail())
-      )
-      .subscribe((res: any) => {
-        this.isLoading = false;
-        this.dataSource = res;
-      });
+    this.search$.next();
   }
-  ngAfterViewInit(): void {}
+ 
   ngOnDestroy(): void {
     this.search$.complete();
     this.unsubscribe$.next();
